+++ src/auth-token-helpers/get-token-using-iam.ts
@@ -20,7 +20,7 @@ export interface IGetTokenUsingIamOpts {
   }
 
   /**
-   * Optional function for injecting credentials instead of loading credentials through the awscred library 
+   * Optional function for injecting credentials instead of loading credentials through the awscred library
    */
   credentialFunc?: () => Promise<IAwsAuth.IAwsCredentials>
 
@@ -155,7 +155,9 @@ export class IamTokenManager {
           return this.awsAuthClient.getTokenUsingIamLogin({
             role: this.role,
             stsRegion: this.opts.stsRegion,
-            credentials: this.opts.credentialFunc ? await this.opts.credentialFunc() : await loadCredentials(),
+            credentials: this.opts.credentialFunc
+              ? await this.opts.credentialFunc()
+              : await loadCredentials(),
             iamRequestHeaders: this.opts.iamRequestHeaders
           })
         },